+++ org.go
@@ -1,30 +1,14 @@
 package mackerel
 
 import (
-	"encoding/json"
-	"net/http"
+	"context"
+
+	v2 "github.com/mackerelio/mackerel-client-go/v2"
 )
 
-// Org information
-type Org struct {
-	Name string `json:"name"`
-}
+type Org = v2.Org
 
 // GetOrg get the org
 func (c *Client) GetOrg() (*Org, error) {
-	req, err := http.NewRequest("GET", c.urlFor("/api/v0/org").String(), nil)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := c.Request(req)
-	defer closeResponse(resp)
-	if err != nil {
-		return nil, err
-	}
-	var data Org
-	err = json.NewDecoder(resp.Body).Decode(&data)
-	if err != nil {
-		return nil, err
-	}
-	return &data, nil
+	return c.OrgService.Get(context.Background())
 }
